pyproject.toml
[tool.black]
    py36 = true
    include = '\.pyi?$'
    exclude = '''
    /(
        \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist

      # The following are specific to Black, you probably don't want those.
      | blib2to3
      | tests/data
    )/'''


[tool.poetry]
name = "tyk_libs"
version = "0.1.0"
description = ""
authors = ["Simon <simon@tyk.io>"]
readme = "README.md"
packages = [{include = "docs"}]


[tool.poetry.dependencies]
python = "^3.11"
semver = "^3.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"
pyhamcrest = "^2.0.4"
behave = "^1.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
